requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
dash==0.32.2
dash-core-components==0.41.0
dash-dangerously-set-inner-html==0.0.1
dash-html-components==0.13.2
dash-renderer==0.14.1
dash-table==3.1.11
dash-table-experiments==0.6.0
decorator==4.3.0
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Compress==1.4.0
gast==0.2.2
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.12.0
idna==2.7
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.1
jsonschema==2.6.0
jupyter-core==4.4.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
nbformat==4.4.0
numpy==1.15.3
pandas==0.23.4
Pillow==5.4.1
plotly==3.3.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pyparsing==2.3.1
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
PyYAML==3.13
requests==2.20.0
retrying==1.3.3
rsa==4.0
singledispatch==3.4.0.3
six==1.11.0
style==1.1.0
termcolor==1.1.0
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
torchvision==0.2.1
traitlets==4.3.2
update==0.0.1
uritemplate==3.0.0
urllib3==1.24
Werkzeug==0.14.1
